+++ frontend/src/domains/wishlist/components/WishForm/index.tsx
@@ -1,5 +1,5 @@
 import Button from '@components/actions/Button';
-import Input from '@components/actions/Input';
+import Input from '@components/actions/Input/Input';
 import ErrorMessage from '@components/errors/ErrorMessage';
 
 import styled from '@emotion/styled';
