+++ frontend/src/pages/CreatePickeatWithLocation.tsx
@@ -1,7 +1,7 @@
 import AddressList from '@domains/pickeat/components/AddressList';
 
 import Button from '@components/actions/Button';
-import Input from '@components/actions/Input';
+import Input from '@components/actions/Input/Input';
 import SearchBar from '@components/actions/SearchBar';
 import Select from '@components/actions/Select';
 import ErrorMessage from '@components/errors/ErrorMessage';
