+++ frontend/src/pages/CreateRoom.tsx
@@ -1,7 +1,7 @@
 import InviteMember from '@domains/room/components/RoomDetailTab/InviteMember';
 
 import Button from '@components/actions/Button';
-import Input from '@components/actions/Input';
+import Input from '@components/actions/Input/Input';
 import ErrorMessage from '@components/errors/ErrorMessage';
 import { HEADER_HEIGHT } from '@components/layouts/Header';
