+++ frontend/src/pages/ProfileInit.tsx
@@ -1,5 +1,5 @@
 import Button from '@components/actions/Button';
-import Input from '@components/actions/Input';
+import Input from '@components/actions/Input/Input';
 import { HEADER_HEIGHT } from '@components/layouts/Header';
 
 import { useAuth } from '@domains/login/context/AuthProvider';
