+++ frontend/src/shared/components/share/CopyLink.tsx
@@ -1,5 +1,5 @@
 import Button from '@components/actions/Button';
-import Input from '@components/actions/Input';
+import Input from '@components/actions/Input/Input';
 
 import { copyLink } from '@utils/copyLink';
